+++ BeamJoyServer/services/config.lua
@@ -87,6 +87,7 @@ local M = {
             amount = 15,
             maxPerPlayer = 1,
             models = { "simple_traffic" },
+            weights = {},
             smartSelection = false,
             parkedAmount = 0,
             parkedMaxPerPlayer = 1,